Fixed targetAddress blank rule for dedicated tenants.
